Renamed project name to django-contact to shorten the project name.
Renamed django-contact-form/ to contact/ to make it easier to load the module.
The chances of this generic name overlapping is small. This breaks
compatability with previous versions of django-contact-form.

Modified the documentation to reflect the above changes.

Removed a lot of documentation from the source code. Basic documentation
remains to give the developer an idea of what or why the class or function
exists. Detailed documentation is kept in the docs/ directory. This makes it
easier to keep versions of the documentation up-to-date.

forms.py: Added option to initialize name and email fields with the current
request's user information.